requirements.txt
python-matter-server==5.2.1
aiohttp==3.9.3
Flask==3.0.2
Jinja2==3.1.3
waitress==3.0.0
swagger-ui-py==23.9.23 
